+++ HISTORY.rst
@@ -28,6 +28,8 @@ latest (unreleased)
 
 **Bugfixes**
 
+* BSQOQ-36: Don't cancel refund when using qoqa_offer_update
+
 **Build**
 
 * Upgrade nginx to 9.0-1.2.1